backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
twilio==8.11.0
openai==1.10.0
celery==5.3.6
redis==5.0.1
python-multipart==0.0.6
pandas==2.2.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
aiofiles==23.2.1
httpx==0.26.0
langchain==0.1.4
langchain-openai==0.0.3
pyotp==2.9.0
qrcode==7.4.2
